backend/requirements.txt
asgiref==3.8.1
black==25.1.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.2.1
Django==4.2.15
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
iniconfig==2.1.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-django==4.11.1
pytest-pythonpath==0.7.3
python-dotenv==1.1.1
requests==2.32.5
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.5.0
